==> traffic_simulation_code/1_create_nodes_and_edges.ipynb <==
the purpose of this code is to run road_link_v3 to create nodes_df and edges_df

this .ipynb was originally named 241216_road_link_sandbox_v2.ipynb

note: I updated the code to run with the latest penalty_dictionary

import sys, os
import glob
code_path = os.path.join(os.path.dirname(os.getcwd()), 'code')
sys.path.append(code_path)
from road_link_v3 import *
pd.set_option('display.max_columns', None)

penalty_dictionary = {
   'Minor-Minor': 0.7,
   'Major-Major': 1.0,
   'Minor-Link': 0.9,
   'Link-Major': 0.9,
   'Link-Link': 0.9,
   'Link-Minor': 0.9,
   'Minor-Major': 0.6,
   'Minor-no_connections': 0.6,
   'Link-no_connections': 0.9,
   'Major-no_connections': 0.9,
   'Major-Link': 0.9,
   'Major-Minor': 0.9
}

nodes_df, edges_df = main_preparation(penalty_dictionary)

# save_path = os.path.join(os.path.dirname(os.getcwd()), 'datasets', 'processed')
# nodes_df.to_file(save_path + "/" + "nodes.gpkg", driver="GPKG")
# edges_df.to_file(save_path + "/" + "edges.gpkg", driver="GPKG")
