==> traffic_simulation_code/2_create_origin_destination_matrix.ipynb <==
the purpose of this code is to create a final origin destination matrix with origin_destination_v2

this file was originally called 241216_od_sandbox_v1.ipynb

import sys, os
import glob
code_path = os.path.join(os.path.dirname(os.getcwd()), 'code')
sys.path.append(code_path)
from origin_destination_v2 import *
pd.set_option('display.max_columns', None)

od_matrix = main_origin_destination_matrix_function()

# save_path = os.path.join(os.path.dirname(os.getcwd()), 'datasets', 'processed', 'od_matrix.csv')
# od_matrix.to_csv(save_path, index=False)
