==> traffic_simulation_code/3_set_weights_for_ffs.ipynb <==
the purpose of this code is to set the correct weights for FFS using setting_weights_v1

this file was originally called 241231_setting_weights_v1.ipynb

import sys, os
import glob
code_path = os.path.join(os.path.dirname(os.getcwd()), 'code')
sys.path.append(code_path)
from setting_weights_v1 import *
pd.set_option('display.max_columns', None)

# free_flow_route_times = main_free_flow_query()

# save_path = os.path.join(os.path.dirname(os.getcwd()), 'datasets', 'processed')
# free_flow_route_times.to_csv(save_path + '/' + 'gmaps_free_flow_route_times.csv', index=False)

create a df of the simulated routes -- this will be a starting point and later I will iterate over the weights

this step uses route_optimization_v2.py and references 241223_optimization_sandbox_v2.ipynb

# from route_optimization_v2 import *
# import ast

# edges_path = read_edges_path()
# edges_df = read_edges(edges_path)

# od_matrix_path = read_od_matrix_path()
# od_matrix = read_od_matrix(od_matrix_path)

# # filter the od_matrix based on the Google Maps routes
# gmaps_path = os.path.join(os.path.dirname(os.getcwd()), 'datasets', 'processed', 'gmaps_free_flow_route_times.csv')
# gmaps_df = pd.read_csv(gmaps_path)
# od_matrix_sample = od_matrix[od_matrix['person_id'].isin(gmaps_df['person_id'])]

# # create the network graph
# G = create_network_graph(edges_df)
# # calculate routes using the pre-built graph
# # vehicles_df is another representation of the OD Matrix data
# vehicles_df, failures_df = calculate_initial_routes(G, od_matrix_sample, edges_df)

# def add_estimated_total_time(vehicles_df, edges_df):
#     #1. create dictionary for quick lookup of free flow times
#     free_flow_times = dict(zip(edges_df['link_id'], edges_df['free_flow_travel_time']))
    
#     def calculate_route_time(route_links):
#         #2. convert string to list if needed
#         if isinstance(route_links, str):
#             route_links = ast.literal_eval(route_links)
#         #3. sum up times for valid route lists
#         if isinstance(route_links, list):
#             return sum(free_flow_times.get(link, 0) for link in route_links)
#         return 0
    
#     #4. apply the calculation to each row
#     vehicles_df['estimated_total_time'] = vehicles_df['route_links'].apply(calculate_route_time)
#     return vehicles_df

# vehicles_df = add_estimated_total_time(vehicles_df, edges_df)
# # save vehicles_df
# save_od_with_routes = os.path.join(os.path.dirname(os.getcwd()), 'datasets', 'processed', 'test_od_with_routes.csv')
# vehicles_df.to_csv(save_od_with_routes, index=False)

testing the setting weights

gmaps_df = read_gmaps_df()
vehicles_df = read_vehicles_df()
# vehicles_df = merge_gmaps_and_vehicles(gmaps_df, vehicles_df)

# penalty_combination_dictionary = {
#    'Minor-Minor': [0.5, 0.8],
#    'Major-Major': [1, 1.3],
#    'Minor-Link': [0.8],
#    'Link-Major': [0.8],
#    'Link-Link': [0.8],
#    'Link-Minor': [0.8],
#    'Minor-Major': [0.5, 0.8],
#    'Minor-no_connections': [0.5, 0.8],
#    'Link-no_connections': [0.8],
#    'Major-no_connections': [0.8],
#    'Major-Link': [0.8],
#    'Major-Minor': [0.8]
# }

# penalty_combination_dictionary = {
#     'Minor-Minor': [1],
#     'Major-Major': [1],
#     'Minor-Link': [1],
#     'Link-Major': [1],
#     'Link-Link': [1],
#     'Link-Minor': [1],
#     'Minor-Major': [1],
#     'Minor-no_connections': [1],
#     'Link-no_connections': [1],
#     'Major-no_connections': [1],
#     'Major-Link': [1],
#     'Major-Minor': [1]
# }

# penalty_combination_dictionary = {
#    'Minor-Minor': [0.6, 0.9],
#    'Major-Major': [1, 1.3],
#    'Minor-Link': [0.9],
#    'Link-Major': [0.9],
#    'Link-Link': [0.9],
#    'Link-Minor': [0.9],
#    'Minor-Major': [0.6, 0.9],
#    'Minor-no_connections': [0.6, 0.9],
#    'Link-no_connections': [0.9],
#    'Major-no_connections': [0.9],
#    'Major-Link': [0.9],
#    'Major-Minor': [0.9],
# }

# this one has too many combinations
# penalty_combination_dictionary = {
#    'Minor-Minor': [0.8, 1.1],
#    'Major-Major': [1.0, 1.3],
#    'Minor-Link': [0.9, 1.1],
#    'Link-Major': [0.9, 1.1],
#    'Link-Link': [0.9, 1.1],
#    'Link-Minor': [0.9, 1.1],
#    'Minor-Major': [0.7, 1.0],
#    'Minor-no_connections': [0.7, 1.0],
#    'Link-no_connections': [1.0],
#    'Major-no_connections': [0.9],
#    'Major-Link': [1.0],
#    'Major-Minor': [0.7, 0.9]
# }

# penalty_combination_dictionary = {
#    'Minor-Minor': [0.8, 1.1],
#    'Major-Major': [1.0, 1.3],
#    'Minor-Link': [1.0],
#    'Link-Major': [1.0],
#    'Link-Link': [1.0],
#    'Link-Minor': [1.0],
#    'Minor-Major': [0.7, 1.0],
#    'Minor-no_connections': [0.85],
#    'Link-no_connections': [1.0],
#    'Major-no_connections': [0.9],
#    'Major-Link': [1.0],
#    'Major-Minor': [0.7, 0.9]
# }

# penalty_combination_dictionary = {
#    'Minor-Minor': [1.1],
#    'Major-Major': [1.2],
#    'Minor-Link': [0.9, 1.1],
#    'Link-Major': [0.9, 1.1],
#    'Link-Link': [0.9, 1.1],
#    'Link-Minor': [0.9, 1.1],
#    'Minor-Major': [0.7, 1.0],
#    'Minor-no_connections': [0.85],
#    'Link-no_connections': [1.0],
#    'Major-no_connections': [0.9],
#    'Major-Link': [1.0],
#    'Major-Minor': [0.7, 0.9]
# }

# penalty_combination_dictionary = {
#    'Minor-Minor': [0.8, 1.1],
#    'Major-Major': [1.2, 1.3],
#    'Minor-Link': [0.8, 1.1],
#    'Link-Major': [1.0],
#    'Link-Link': [1.1],
#    'Link-Minor': [0.9],
#    'Minor-Major': [0.7, 1],
#    'Minor-no_connections': [0.8],
#    'Link-no_connections': [0.9],
#    'Major-no_connections': [0.9],
#    'Major-Link': [0.9, 1.1],
#    'Major-Minor': [0.6, 0.7]
# }

penalty_combination_df = create_penalty_combination_df(penalty_combination_dictionary)
print(penalty_combination_df.shape[0])
test = iterate_and_measure_penalty_success(vehicles_df, gmaps_df, penalty_combination_df, minutes_list=[50, 30, 10, 5, 3])

test[test['num_within_10_min'] == test['num_within_10_min'].max()].sort_values(by='num_within_5_min', ascending=False).head(10)

test[test['num_within_3_min'] == test['num_within_3_min'].max()]
